Squash-merge: platinum remaster.
- removed custom logger classes: switched to log4j.
- update on maven assemblies.
- ptk-zipdist now creates actual distribution.
- cleanup, comments, deletion of stale code + formatting.
- removed obsolete lobo module (was broken).
- UI tweaks, fixed fonts, nimbus L&F,etc.

+++ ptk-web/src/main/java/nl/esciencecenter/ptk/web/content/ByteBufferBody.java
@@ -2,7 +2,7 @@
  * Copyright 2012-2014 Netherlands eScience Center.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
- * You may not use this file except in compliance with the License. 
+ * You may not use this file except in compliance with the License.
  * You may obtain a copy of the License at the following location:
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
@@ -12,7 +12,7 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- * 
+ *
  * For the full license, see: LICENSE.txt (located in the root folder of this distribution).
  * ---
  */
@@ -20,13 +20,12 @@
 
 package nl.esciencecenter.ptk.web.content;
 
-import java.io.IOException;
-import java.io.OutputStream;
-
 import nl.esciencecenter.ptk.web.PutMonitor;
-
 import org.apache.http.entity.mime.MIME;
 import org.apache.http.entity.mime.content.AbstractContentBody;
+
+import java.io.IOException;
+import java.io.OutputStream;
 
 /**
  * Byte Buffer body which can be monitored during the upload.
@@ -40,7 +39,7 @@ public class ByteBufferBody extends AbstractContentBody {
     private PutMonitor putMonitor;
 
     public ByteBufferBody(final byte[] data, final String mimeType, final String filename,
-            PutMonitor optPutMonitor) {
+                          PutMonitor optPutMonitor) {
         super(mimeType);
         if (data == null) {
             throw new IllegalArgumentException("byte[] may not be null");
@@ -70,7 +69,7 @@ public class ByteBufferBody extends AbstractContentBody {
                 numToWrite = chunkSize;
             }
 
-            out.write(data, (int) numWritten, (int) numToWrite);
+            out.write(data, numWritten, numToWrite);
             numWritten += numToWrite;
             if (putMonitor != null) {
                 putMonitor.bytesWritten(numWritten);
